# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Perto de Mim API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProfessionalLevel = typeof ProfessionalLevel[keyof typeof ProfessionalLevel];


export const ProfessionalLevel = {
  amador: 'amador',
  profissional: 'profissional',
} as const;

export interface Professional {
  id: number;
  name: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  profession: string;
  skills?: string[] | null;
  /** @nullable */
  professionDetail?: string | null;
  /** @nullable */
  lessonType?: string | null;
  level: ProfessionalLevel;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  createdAt: string;
}

export type ProfessionalInputLevel = typeof ProfessionalInputLevel[keyof typeof ProfessionalInputLevel];


export const ProfessionalInputLevel = {
  amador: 'amador',
  profissional: 'profissional',
} as const;

export interface ProfessionalInput {
  name: string;
  photoUrl?: string;
  linkUrl?: string;
  profession?: string;
  skills: string[];
  professionDetail?: string;
  lessonType?: string;
  level: ProfessionalInputLevel;
  lat?: number;
  lng?: number;
  address?: string;
  phone?: string;
}

export interface Review {
  id: number;
  professionalId: number;
  reviewerName: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface ReviewInput {
  reviewerName?: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment?: string;
}

export interface PlatformStats {
  totalProfessionals: number;
  totalProfessions: number;
  recentlyAdded?: number;
}

export type ListProfessionalsParams = {
/**
 * Search query
 */
query?: string;
/**
 * Latitude for proximity search
 */
lat?: number;
/**
 * Longitude for proximity search
 */
lng?: number;
/**
 * Search radius in km (default 1)
 */
radiusKm?: number;
};