# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Perto de Mim API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: professionals
    description: Professional profiles
  - name: reviews
    description: Professional reviews
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /professionals:
    get:
      operationId: listProfessionals
      tags: [professionals]
      summary: List all registered professionals
      parameters:
        - name: query
          in: query
          schema:
            type: string
          description: Search query
        - name: lat
          in: query
          schema:
            type: number
          description: Latitude for proximity search
        - name: lng
          in: query
          schema:
            type: number
          description: Longitude for proximity search
        - name: radiusKm
          in: query
          schema:
            type: number
          description: Search radius in km (default 1)
      responses:
        "200":
          description: List of professionals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Professional"
    post:
      operationId: createProfessional
      tags: [professionals]
      summary: Register a new professional
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfessionalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professional"

  /professionals/stats:
    get:
      operationId: getProfessionalsStats
      tags: [professionals]
      summary: Get platform stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformStats"

  /professionals/{id}:
    get:
      operationId: getProfessional
      tags: [professionals]
      summary: Get a professional by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Professional found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professional"
        "404":
          description: Not found
    put:
      operationId: updateProfessional
      tags: [professionals]
      summary: Update a professional profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfessionalInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professional"
        "404":
          description: Not found

  /professionals/{id}/reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews for a professional
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review for a professional
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Invalid data
        "404":
          description: Professional not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Professional:
      type: object
      required: [id, name, profession, level, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        photoUrl:
          type: ["string", "null"]
        linkUrl:
          type: ["string", "null"]
        profession:
          type: string
        skills:
          type: array
          items:
            type: string
          nullable: true
        professionDetail:
          type: ["string", "null"]
        lessonType:
          type: ["string", "null"]
        level:
          type: string
          enum: [amador, profissional]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string

    ProfessionalInput:
      type: object
      required: [name, skills, level]
      properties:
        name:
          type: string
        photoUrl:
          type: string
        linkUrl:
          type: string
        profession:
          type: string
        skills:
          type: array
          items:
            type: string
        professionDetail:
          type: string
        lessonType:
          type: string
        level:
          type: string
          enum: [amador, profissional]
        lat:
          type: number
        lng:
          type: number
        address:
          type: string
        phone:
          type: string

    Review:
      type: object
      required: [id, professionalId, reviewerName, rating, createdAt]
      properties:
        id:
          type: integer
        professionalId:
          type: integer
        reviewerName:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: ["string", "null"]
        createdAt:
          type: string

    ReviewInput:
      type: object
      required: [rating]
      properties:
        reviewerName:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string

    PlatformStats:
      type: object
      required: [totalProfessionals, totalProfessions]
      properties:
        totalProfessionals:
          type: integer
        totalProfessions:
          type: integer
        recentlyAdded:
          type: integer