# lib/api-zod/src/generated/types/platformStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Perto de Mim API
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformStats {
  totalProfessionals: number;
  totalProfessions: number;
  recentlyAdded?: number;
}