# lib/api-zod/src/generated/types/professional.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Perto de Mim API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfessionalLevel } from './professionalLevel';

export interface Professional {
  id: number;
  name: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  profession: string;
  skills?: string[] | null;
  /** @nullable */
  professionDetail?: string | null;
  /** @nullable */
  lessonType?: string | null;
  level: ProfessionalLevel;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  createdAt: string;
}