# lib/api-zod/src/generated/types/professionalInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Perto de Mim API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfessionalInputLevel } from './professionalInputLevel';

export interface ProfessionalInput {
  name: string;
  photoUrl?: string;
  linkUrl?: string;
  profession?: string;
  skills: string[];
  professionDetail?: string;
  lessonType?: string;
  level: ProfessionalInputLevel;
  lat?: number;
  lng?: number;
  address?: string;
  phone?: string;
}