# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Perto de Mim API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  professionalId: number;
  reviewerName: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}